Projeto Funcionando
-Arrumado erro de malloc no bytizador
-Mudado com a altura e o lixo são guardados e lidos

+++ Compactador/FilaArvore.h
@@ -8,7 +8,7 @@ typedef struct NoFilAr
     struct NoFilAr* prox;
     No* dado;
     char* cod;
-    short int indice;
+    int indice;
     char h;
 } NoFilAr;
